requirements/requirements.txt
alembic==1.8.1
amqp==5.1.1
aniso8601==9.0.1
appdirs==1.4.4
asgiref==3.5.2
attrs==22.1.0
Automat==20.2.0
automaton==3.0.1
autopage==0.5.1
Babel==2.8.0
bcrypt==4.0.0
blinker==1.4
boto3==1.24.89
botocore==1.27.89
cachetools==5.2.0
castellan==4.0.0
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
-e git+https://opendev.org/openstack/cinder.git@dcec2f6f01ffc63dc058f641370e9f5a0bad07e4#egg=cinder
click==8.1.3
cliff==4.0.0
cloud-init==22.3.4
cmd2==2.4.2
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
configshell-fb==1.1.28
constantly==15.1.0
cryptography==38.0.2
cursive==0.2.2
dbcounter @ file:///home/wonjoon/devstack/tools/dbcounter
dbus-python==1.2.18
debtcollector==2.5.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
Django==4.1.3
django-environ==0.9.0
djangorestframework==3.14.0
dnspython==2.2.1
dogpile.cache==1.1.8
elementpath==3.0.2
etcd3==0.12.0
etcd3gw==2.1.0
eventlet==0.33.1
extras==1.0.0
fasteners==0.18
filelock==3.8.0
fixtures==4.0.1
Flask==2.2.2
Flask-RESTful==0.3.9
futurist==2.4.1
-e git+https://opendev.org/openstack/glance.git@199722a65a88f9c28b554b180ec36415ed3a51ac#egg=glance
glance-store==4.1.0
google-api-core==2.10.2
google-api-python-client==2.64.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
greenlet==1.1.3
grpcio==1.49.1
httplib2==0.20.4
hyperlink==21.0.0
idna==3.4
importlib-metadata==5.0.0
incremental==21.3.0
iso8601==1.1.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.16.0
keyring==23.5.0
-e git+https://opendev.org/openstack/keystone.git@7d4047cb69139c22f4e321e9ba8b287a47168df2#egg=keystone
keystoneauth1==5.0.0
keystonemiddleware==10.1.0
kombu==5.2.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libvirt-python==8.0.0
logutils==0.3.5
lxml==4.9.1
Mako==1.2.3
MarkupSafe==2.1.1
microversion-parse==1.0.1
more-itertools==8.10.0
msgpack==1.0.4
munch==2.5.0
mysqlclient==1.4.6
ncclient==0.6.13
netaddr==0.8.0
netifaces==0.11.0
networkx==2.8.7
-e git+https://opendev.org/openstack/neutron.git@a358bb802bb99b6c89e0acfe0a77ad45612c825e#egg=neutron
neutron-lib==3.1.0
-e git+https://opendev.org/openstack/nova.git@2eb358cdcec36fcfe5388ce6982d2961ca949d0a#egg=nova
numpy==1.23.3
oauth2client==4.1.3
oauthlib==3.2.1
-e git+https://opendev.org/openstack/placement.git@4e345e55c2d695f7c271ac5119cfed761eb7d206#egg=openstack_placement
openstacksdk==0.102.0
os-brick==6.1.0
os-client-config==2.1.0
os-ken==2.5.0
os-resource-classes==1.1.0
os-service-types==1.7.0
os-testr==3.0.0
os-traits==2.9.0
os-vif==2.8.0
os-win==5.7.1
osc-lib==2.6.2
osc-placement==4.0.0
oslo.cache==3.1.0
oslo.concurrency==5.0.1
oslo.config==9.0.0
oslo.context==5.0.0
oslo.db==12.2.0
oslo.i18n==5.1.0
oslo.limit==2.0.1
oslo.log==5.0.0
oslo.messaging==14.0.0
oslo.metrics==0.5.0
oslo.middleware==5.0.0
oslo.policy==4.0.0
oslo.privsep==3.0.1
oslo.reports==2.4.0
oslo.rootwrap==6.3.1
oslo.serialization==5.0.0
oslo.service==3.0.0
oslo.upgradecheck==2.0.0
oslo.utils==6.0.1
oslo.versionedobjects==3.0.1
oslo.vmware==4.0.1
osprofiler==3.4.3
ovs==2.17.1.post1
ovsdbapp==2.1.0
packaging==21.3
paramiko==2.11.0
passlib==1.7.4
Paste==3.5.2
PasteDeploy==2.1.1
pbr==5.11.0
pcp==5.0
pecan==1.4.2
pexpect==4.8.0
platformdirs==2.5.2
pluggy==1.0.0
prettytable==3.4.1
prometheus-client==0.14.1
protobuf==4.21.7
psutil==5.9.2
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycadf==3.1.1
pycparser==2.21
pydot==1.4.2
Pygments==2.11.2
PyGObject==3.42.1
PyHamcrest==2.0.2
pyinotify==0.9.6
PyJWT==2.5.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyperclip==1.8.2
pyroute2==0.7.3
pyrsistent==0.18.1
pysaml2==7.2.1
pyserial==3.5
python-apt==2.3.0+ubuntu2.1
python-barbicanclient==5.4.0
python-cinderclient==9.1.0
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-designateclient==5.0.0
python-glanceclient==4.1.0
python-keystoneclient==5.0.1
python-magic==0.4.24
python-memcached==1.59
python-neutronclient==8.1.0
python-novaclient==18.1.0
python-openstackclient==6.0.0
python-subunit==1.4.0
python-swiftclient==4.1.0
pytz==2022.4
pyudev==0.24.0
PyYAML==6.0
repoze.lru==0.7
requests==2.28.1
requestsexceptions==1.4.0
retrying==1.3.3
rfc3986==1.5.0
Routes==2.5.1
rsa==4.9
rtslib-fb==2.1.75
s3transfer==0.6.0
scrypt==0.8.20
SecretStorage==3.3.1
service-identity==18.1.0
setproctitle==1.3.2
simplegeneric==0.8.1
simplejson==3.17.6
six==1.16.0
sortedcontainers==2.4.0
sos==4.4
SQLAlchemy==1.4.41
sqlalchemy-migrate==0.13.0
sqlparse==0.4.3
ssh-import-id==5.11
statsd==3.3.0
stevedore==4.1.1
suds-community==1.1.2
systemd-python==234
tabulate==0.9.0
taskflow==5.0.0
Tempita==0.5.2
tenacity==6.3.1
testresources==2.0.1
testscenarios==0.5.0
testtools==2.5.0
tomli==2.0.1
tooz==3.2.0
tox==3.27.1
Twisted==22.1.0
ubuntu-advantage-tools==27.11.3
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.12
urwid==2.1.2
vine==5.0.0
virtualenv==20.16.5
voluptuous==0.13.1
wadllib==1.3.6
warlock==2.0.1
wcwidth==0.2.5
WebOb==1.8.7
websockify==0.10.0
Werkzeug==2.2.2
wrapt==1.14.1
WSME==0.11.0
xkit==0.0.0
xmlschema==2.1.1
yappi==1.3.6
zipp==3.8.1
zope.interface==5.4.0
zstd==1.5.2.6
